gcc compiler required

+++ tests/posix/sendmsg/client.c
@@ -82,7 +82,7 @@ void run_client(uint16_t port)
         /* Send a message. */
         {
             msg_send.msg_iov = (struct iovec*)iov;
-            msg_send.msg_iovlen = iovlen;
+            msg_send.msg_iovlen = (int)iovlen;
 
             if (sendmsg(sd, &msg_send, 0) != total_iov_size)
                 OE_TEST("sendmsg() failed" == NULL);
@@ -101,7 +101,7 @@ void run_client(uint16_t port)
 
             memset(&msg_recv, 0, sizeof(msg_recv));
             msg_recv.msg_iov = iov_buf;
-            msg_recv.msg_iovlen = iovlen;
+            msg_recv.msg_iovlen = (int)iovlen;
 
             if (recvmsg(sd, &msg_recv, 0) != total_iov_size)
                 OE_TEST("recvmsg() failed" == NULL);
@@ -140,7 +140,7 @@ void run_client(uint16_t port)
         memset(&msg, 0, sizeof(msg));
 
         msg.msg_iov = (struct iovec*)iov;
-        msg.msg_iovlen = iovlen;
+        msg.msg_iovlen = (int)iovlen;
 
         ssize_t m = sendmsg(sd, &msg, 0);
